fix: ensure that global Pdr object has frames
Ensure that the global Pdr object has as many frames as the Bmc problem.
Currently, it was possible that if nothing is to be added to the global
frame (i.e., F_0 = True), no Solver object is created as well. This
violates invariants of much of the rest of the code.